+++ src/test/java/one/util/streamex/UnknownSizeSpliteratorTest.java
@@ -97,37 +97,4 @@ public class UnknownSizeSpliteratorTest {
         assertEquals(12497500, StreamSupport.stream(new UnknownSizeSpliterator.USOfRef<>(input.iterator()), true)
                 .mapToInt(x -> x).sum());
     }
-
-    @Test
-    public void testOptimize() {
-        Stream<String> stream = Stream.of("a", "b");
-        assertSame(stream, UnknownSizeSpliterator.optimize(stream));
-        stream = StreamSupport.stream(
-            Spliterators.spliteratorUnknownSize(Arrays.asList("a", "b").iterator(), Spliterator.ORDERED), false)
-                .distinct(); // not Head
-        assertSame(stream, UnknownSizeSpliterator.optimize(stream));
-        stream = StreamSupport.stream(Spliterators.spliterator(Arrays.asList("a", "b").iterator(), 2,
-            Spliterator.ORDERED), false); // SIZED
-        assertSame(stream, UnknownSizeSpliterator.optimize(stream));
-        stream = new ConcurrentLinkedDeque<String>().stream(); // not SIZED
-        assertSame(stream, UnknownSizeSpliterator.optimize(stream));
-
-        AtomicBoolean flag = new AtomicBoolean();
-        stream = StreamSupport.stream(
-            Spliterators.spliteratorUnknownSize(Arrays.asList("a", "b").iterator(), Spliterator.ORDERED), false)
-                .onClose(() -> flag.set(true));
-        Stream<String> optimized = UnknownSizeSpliterator.optimize(stream);
-        assertNotSame(stream, optimized);
-        assertTrue(optimized.spliterator() instanceof UnknownSizeSpliterator);
-        assertFalse(optimized.isParallel());
-        optimized.close();
-        assertTrue(flag.get());
-
-        stream = StreamSupport.stream(Spliterators.spliteratorUnknownSize(Arrays.asList("a", "b").iterator(),
-            Spliterator.ORDERED), true);
-        optimized = UnknownSizeSpliterator.optimize(stream);
-        assertNotSame(stream, optimized);
-        assertTrue(optimized.spliterator() instanceof UnknownSizeSpliterator);
-        assertTrue(optimized.isParallel());
-    }
 }